moveTo, lineTo and curveTo commands on GeneralPath only accept floats in Java 1.5.

+++ src/net/nodebox/graphics/BezierPath.java
@@ -135,15 +135,17 @@ public class BezierPath implements Grob {
         for (PathElement el : elements) {
             switch (el.getCommand()) {
                 case PathElement.COMMAND_MOVETO:
-                    gp.moveTo(el.getX(), el.getY());
+                    gp.moveTo((float)el.getX(), (float)el.getY());
                     break;
                 case PathElement.COMMAND_LINETO:
-                    gp.lineTo(el.getX(), el.getY());
+                    gp.lineTo((float)el.getX(), (float)el.getY());
                     break;
                 case PathElement.COMMAND_CURVETO:
                     c1 = el.getControl1();
                     c2 = el.getControl2();
-                    gp.curveTo(el.getX(), el.getY(), c1.getX(), c1.getY(), c2.getX(), c2.getY());
+                    gp.curveTo((float)el.getX(), (float)el.getY(),
+                            (float)c1.getX(), (float)c1.getY(),
+                            (float)c2.getX(), (float)c2.getY());
                     break;
                 case PathElement.COMMAND_CLOSE:
                     gp.closePath();
